CampaignDropdown
Background: Added property to app called ChosenCampaign and assigned it to the value of the added dropdown menu, which includes the names of the five campaigns.

Summary: (1) Adding campaign as a parameter to functions tells the app which sheet in the spreadsheet to open and interact with. (2) The addition of the START_ROW constant as well as minor modifications to getStartRow (server), getEndRow (server), Next (client), and Previous (client) functions ensure that callers/texters are only shown the people they were assigned.